python/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
distro==1.9.0
fastapi==0.116.1
grpcio==1.74.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jiter==0.10.0
numpy==2.3.2
openai==1.98.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
portalocker==3.2.0
protobuf==6.31.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pypdfium2==4.30.0
python-dotenv==1.1.1
qdrant-client==1.15.1
sniffio==1.3.1
soupsieve==2.7
starlette==0.47.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
